+++ bot/src/test/java/edu/java/hw1/ListHandlerTest.java
@@ -1,4 +1,4 @@
-package edu.hw1;
+package edu.java.hw1;
 
 import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.model.Message;
@@ -9,11 +9,12 @@ import edu.java.bot.model.Bot;
 import edu.java.bot.model.BotUser;
 import edu.java.bot.model.Chat;
 import edu.java.bot.repository.CommandName;
-import edu.java.bot.service.NoCommandHandler;
+import edu.java.bot.service.ListHandler;
 import edu.java.bot.service.UserMessageHandler;
 import edu.java.bot.service.UserMessageHandlerImpl;
-import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
@@ -21,7 +22,7 @@ import org.mockito.Mockito;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.mockito.Mockito.mock;
 
-public class NoCommandHandlerTest {
+public class ListHandlerTest {
     @Mock TelegramBot telegramBot = new TelegramBot("12345");
     HashMap<BotUser, CommandName> isWaiting = new HashMap<>();
 
@@ -39,25 +40,33 @@ public class NoCommandHandlerTest {
         "12345",
         "aa",
         "1",
-        "2",
-        "sorry",
-        "3",
-        "7",
-        "8",
-        "4 ",
-        "^(https?://){1}([\\w\\Q$-_+!*'(),%\\E]+\\.)+(\\w{2,63})(:\\d{1,4})?([\\w\\Q/$-_+!*'(),%\\E]+\\.?[\\w\\Q$-_+!*'(),%\\E={0-5}?&.])*/?$",
-        "6"
+        "1",
+        "1",
+        "1",
+        "Here you are: ",
+        "You have no links being tracked. Print /track to add a link",
+        "1",
+        "1",
+        "1"
 
     );
 
     @Test
-    void processesUnAuthorized() {
-        Bot bot = new Bot(
-            telegramBot,
-            Map.of(),
-            Map.of()
+    void appliesListCommand() {
+        Chat chat1 = new Chat(
+            botUser.chatId(),
+            botUser.id(),
+            botUser.name(),
+            List.of("https://stackoverflow.com/search?q=unsupported%20link")
         );
-        var handler = new NoCommandHandler(applicationConfig, true);
+        isWaiting.put(botUser, null);
+        var bot = new Bot(
+            telegramBot,
+            Map.of(botUser, chat1),
+            isWaiting
+        );
+
+        var handler = new ListHandler(applicationConfig, true);
         Mockito.when(update.message()).thenReturn(message);
         Mockito.when(message.text()).thenReturn(CommandName.LIST.getCommand());
         Mockito.when(message.chat()).thenReturn(chat);
@@ -67,27 +76,28 @@ public class NoCommandHandlerTest {
         var result = handler.handle(bot, messageHandler, update);
 
         assertThat(result.getParameters().get("text")).isEqualTo(
-            ("sorry")
-        );
+            ("Here you are: " + Arrays.deepToString(bot.chats().get(botUser).links().toArray())
+            ));
     }
 
     @Test
-    void startsTracking() {
+    void notifiesWhenNoLinks() {
         Chat chat1 = new Chat(
             botUser.chatId(),
             botUser.id(),
             botUser.name(),
-            new ArrayList<>()
+            List.of()
         );
-        isWaiting.put(botUser, CommandName.TRACK);
+        isWaiting.put(botUser, null);
         Bot bot = new Bot(
             telegramBot,
             Map.of(botUser, chat1),
             isWaiting
         );
-        var handler = new NoCommandHandler(applicationConfig, true);
+        var handler = new ListHandler(applicationConfig, true);
+
         Mockito.when(update.message()).thenReturn(message);
-        Mockito.when(message.text()).thenReturn("https://stackoverflow.com/search?q=unsupported%20link");
+        Mockito.when(message.text()).thenReturn(CommandName.LIST.getCommand());
         Mockito.when(message.chat()).thenReturn(chat);
         Mockito.when(message.chat().id()).thenReturn(1L);
         Mockito.when(message.from()).thenReturn(user);
@@ -95,28 +105,21 @@ public class NoCommandHandlerTest {
         var result = handler.handle(bot, messageHandler, update);
 
         assertThat(result.getParameters().get("text")).isEqualTo(
-            ("4 https://stackoverflow.com/search?q=unsupported%20link")
+            ("You have no links being tracked. Print /track to add a link")
         );
     }
 
     @Test
-    void startsUnTracking() {
-        Chat chat1 = new Chat(
-            botUser.chatId(),
-            botUser.id(),
-            botUser.name(),
-            new ArrayList<>()
-        );
-        chat1.links().add("https://stackoverflow.com/search?q=unsupported%20link");
-        isWaiting.put(botUser, CommandName.UNTRACK);
+    void asksToRegister() {
         Bot bot = new Bot(
             telegramBot,
-            Map.of(botUser, chat1),
-            isWaiting
+            Map.of(),
+            Map.of()
         );
-        var handler = new NoCommandHandler(applicationConfig, true);
+        var handler = new ListHandler(applicationConfig, true);
+
         Mockito.when(update.message()).thenReturn(message);
-        Mockito.when(message.text()).thenReturn("https://stackoverflow.com/search?q=unsupported%20link");
+        Mockito.when(message.text()).thenReturn(CommandName.LIST.getCommand());
         Mockito.when(message.chat()).thenReturn(chat);
         Mockito.when(message.chat().id()).thenReturn(1L);
         Mockito.when(message.from()).thenReturn(user);
@@ -124,36 +127,8 @@ public class NoCommandHandlerTest {
         var result = handler.handle(bot, messageHandler, update);
 
         assertThat(result.getParameters().get("text")).isEqualTo(
-            ("4 []")
-        );
-        assertThat(chat1.links()).isEqualTo(new ArrayList<>());
-    }
-
-    @Test
-    void badLink() {
-        Chat chat1 = new Chat(
-            botUser.chatId(),
-            botUser.id(),
-            botUser.name(),
-            new ArrayList<>()
-        );
-        isWaiting.put(botUser, CommandName.TRACK);
-        Bot bot = new Bot(
-            telegramBot,
-            Map.of(botUser, chat1),
-            isWaiting
-        );
-        var handler = new NoCommandHandler(applicationConfig, true);
-        Mockito.when(update.message()).thenReturn(message);
-        Mockito.when(message.text()).thenReturn("stackoverflow.com/search?q=unsupported%20link");
-        Mockito.when(message.chat()).thenReturn(chat);
-        Mockito.when(message.chat().id()).thenReturn(1L);
-        Mockito.when(message.from()).thenReturn(user);
-
-        var result = handler.handle(bot, messageHandler, update);
-
-        assertThat(result.getParameters().get("text")).isEqualTo(
-            ("sorry")
+            ("aa")
         );
     }
 }
+
