+++ bot/src/test/java/edu/java/hw1/StartHandlerTest.java
@@ -1,4 +1,4 @@
-package edu.hw1;
+package edu.java.hw1;
 
 import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.model.Message;
@@ -9,10 +9,9 @@ import edu.java.bot.model.Bot;
 import edu.java.bot.model.BotUser;
 import edu.java.bot.model.Chat;
 import edu.java.bot.repository.CommandName;
-import edu.java.bot.service.ListHandler;
+import edu.java.bot.service.StartHandler;
 import edu.java.bot.service.UserMessageHandler;
 import edu.java.bot.service.UserMessageHandlerImpl;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -22,12 +21,11 @@ import org.mockito.Mockito;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.mockito.Mockito.mock;
 
-public class ListHandlerTest {
+public class StartHandlerTest {
     @Mock TelegramBot telegramBot = new TelegramBot("12345");
-    HashMap<BotUser, CommandName> isWaiting = new HashMap<>();
 
     Update update = mock(Update.class);
-
+    HashMap<BotUser, CommandName> isWaiting = new HashMap<>();
     Message message = mock(Message.class);
 
     User user = new User(1L);
@@ -36,39 +34,30 @@ public class ListHandlerTest {
     com.pengrad.telegrambot.model.Chat chat = mock(com.pengrad.telegrambot.model.Chat.class);
     BotUser botUser = new BotUser(1L, 1L, null, true);
 
-    ApplicationConfig applicationConfig = new ApplicationConfig(
-        "12345",
-        "aa",
-        "1",
-        "1",
-        "1",
-        "1",
-        "Here you are: ",
-        "You have no links being tracked. Print /track to add a link",
-        "1",
-        "1",
-        "1"
-
-    );
-
     @Test
-    void appliesListCommand() {
-        Chat chat1 = new Chat(
-            botUser.chatId(),
-            botUser.id(),
-            botUser.name(),
-            List.of("https://stackoverflow.com/search?q=unsupported%20link")
-        );
-        isWaiting.put(botUser, null);
-        var bot = new Bot(
+    void registerNewUser() {
+        Bot bot = new Bot(
             telegramBot,
-            Map.of(botUser, chat1),
-            isWaiting
+            new HashMap<>(),
+            new HashMap<>()
         );
+        ApplicationConfig applicationConfig = new ApplicationConfig(
+            "12345",
+            "1",
+            "Registered! Hello, dear ",
+            "123123",
+            "1",
+            "1",
+            "1",
+            "1",
+            "1",
+            "1",
+            "1"
 
-        var handler = new ListHandler(applicationConfig, true);
+        );
+        var handler = new StartHandler(applicationConfig, true);
         Mockito.when(update.message()).thenReturn(message);
-        Mockito.when(message.text()).thenReturn(CommandName.LIST.getCommand());
+        Mockito.when(message.text()).thenReturn(CommandName.START.getCommand());
         Mockito.when(message.chat()).thenReturn(chat);
         Mockito.when(message.chat().id()).thenReturn(1L);
         Mockito.when(message.from()).thenReturn(user);
@@ -76,12 +65,12 @@ public class ListHandlerTest {
         var result = handler.handle(bot, messageHandler, update);
 
         assertThat(result.getParameters().get("text")).isEqualTo(
-            ("Here you are: " + Arrays.deepToString(bot.chats().get(botUser).links().toArray())
-            ));
+            ("Registered! Hello, dear " + botUser.name())
+        );
     }
 
     @Test
-    void notifiesWhenNoLinks() {
+    void saysIsAlreadyRegistered() {
         Chat chat1 = new Chat(
             botUser.chatId(),
             botUser.id(),
@@ -94,10 +83,23 @@ public class ListHandlerTest {
             Map.of(botUser, chat1),
             isWaiting
         );
-        var handler = new ListHandler(applicationConfig, true);
+        ApplicationConfig applicationConfig = new ApplicationConfig(
+            "12345",
+            "1",
+            "1",
+            "1234567",
+            "1",
+            "1",
+            "1",
+            "1",
+            "1",
+            "1",
+            "1"
 
+        );
+        var handler = new StartHandler(applicationConfig, true);
         Mockito.when(update.message()).thenReturn(message);
-        Mockito.when(message.text()).thenReturn(CommandName.LIST.getCommand());
+        Mockito.when(message.text()).thenReturn(CommandName.START.getCommand());
         Mockito.when(message.chat()).thenReturn(chat);
         Mockito.when(message.chat().id()).thenReturn(1L);
         Mockito.when(message.from()).thenReturn(user);
@@ -105,30 +107,7 @@ public class ListHandlerTest {
         var result = handler.handle(bot, messageHandler, update);
 
         assertThat(result.getParameters().get("text")).isEqualTo(
-            ("You have no links being tracked. Print /track to add a link")
-        );
-    }
-
-    @Test
-    void asksToRegister() {
-        Bot bot = new Bot(
-            telegramBot,
-            Map.of(),
-            Map.of()
-        );
-        var handler = new ListHandler(applicationConfig, true);
-
-        Mockito.when(update.message()).thenReturn(message);
-        Mockito.when(message.text()).thenReturn(CommandName.LIST.getCommand());
-        Mockito.when(message.chat()).thenReturn(chat);
-        Mockito.when(message.chat().id()).thenReturn(1L);
-        Mockito.when(message.from()).thenReturn(user);
-
-        var result = handler.handle(bot, messageHandler, update);
-
-        assertThat(result.getParameters().get("text")).isEqualTo(
-            ("aa")
+            ("1234567")
         );
     }
 }
-
